toptrendsApi/toptrendsBgAPI/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
db-dtypes==1.1.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
numpy==1.26.1
packaging==23.2
pandas==2.1.3
proto-plus==1.22.3
protobuf==4.25.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
